refactor: ui redo, part 1

+++ templates/sub.func.go
@@ -3,8 +3,8 @@ package templates
 
 import "fmt"
 
-func(a... int) int {
-	var x int = a[0]
+func(a... int64) int64 {
+	var x int64 = a[0]
 	for i, item := range a {
 		if i == 0 { continue }
 		x = x - item
